+++ src/Array.ts
@@ -1,5 +1,3 @@
-import { map } from "./Object"
-
 /**
  * Remove all elements from the given array that return false from the given
  * predicate `filter`.
@@ -14,39 +12,4 @@ export function filterInPlace<T>(arr: T[], filter: (t: T) => boolean): T[] {
   })
   arr.length = j
   return arr
-}
-
-export function flatten<T>(arr: (T | T[])[], result: T[] = []): T[] {
-  arr.forEach((ea) =>
-    Array.isArray(ea) ? result.push(...ea) : result.push(ea)
-  )
-  return result
-}
-
-export function sortNumeric(arr: number[]): number[] {
-  return arr.sort((a, b) => a - b)
-}
-
-export function rrFindResult<T>(
-  arr: T[],
-  startIndex: number,
-  predicate: (t: T, arrIdx: number, iter: number) => boolean
-): undefined | { result: T; index: number } {
-  for (let iter = 0; iter < arr.length; iter++) {
-    const arrIdx = (iter + startIndex) % arr.length
-    const t = arr[arrIdx]
-    if (predicate(t, arrIdx, iter)) return { result: t, index: arrIdx }
-  }
-  return
-}
-
-/**
- * Treat an array as a round-robin list, starting from `startIndex`.
- */
-export function rrFind<T>(
-  arr: T[],
-  startIndex: number,
-  predicate: (t: T, arrIdx: number, iter: number) => boolean
-): undefined | T {
-  return map(rrFindResult(arr, startIndex, predicate), (ea) => ea.result)
 }